fix: refactor remote index export
