+++ src/pages/Howto/Category/CategoriesSelect.tsx
@@ -1,7 +1,7 @@
 import { observer } from 'mobx-react'
 import { FieldContainer } from '../../../common/Form/FieldContainer'
 import { Select } from 'oa-components'
-import { useCommonStores } from 'src/index'
+import { useCommonStores } from 'src/App'
 import type { ICategory } from 'src/models/categories.model'
 
 export const CategoriesSelect = observer(
