+++ src/pages/Research/Content/ResearchComments/ResearchComments.tsx
@@ -1,7 +1,7 @@
 import { Button, CommentList, CreateComment } from 'oa-components'
 import { useState } from 'react'
 import { MAX_COMMENT_LENGTH } from '../../../../constants'
-import { useCommonStores } from '../../../../'
+import { useCommonStores } from '../../../../App'
 import { logger } from 'src/logger'
 import { useResearchStore } from '../../../../stores/Research/research.store'
 import { Box, Flex } from 'theme-ui'
