+++ src/pages/User/content/index.tsx
@@ -3,7 +3,7 @@ import type { RouteComponentProps } from 'react-router'
 import { Loader } from 'oa-components'
 import { Text } from 'theme-ui'
 import { observer } from 'mobx-react-lite'
-import { useCommonStores } from 'src/index'
+import { useCommonStores } from 'src/App'
 import { ProfileType } from 'src/modules/profile/types'
 import { MemberProfile } from './MemberProfile'
 import { SpaceProfile } from './SpaceProfile'
